Backened/requirements.txt
# ---------------- Core Backend ----------------
pydantic==1.10.13
fastapi==0.95.2
uvicorn==0.20.0
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.2

# ---------------- Vector DB ----------------
# 2) Reinstall pinned compatible versions
chromadb==0.4.24
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# DO NOT USE: langchain_chroma   (remove this completely)
# It breaks older Chroma

# ---------------- LLM Providers ----------------
langchain-google-genai==0.0.11
google-generativeai==0.4.1

# ---------------- Embeddings ----------------
sentence-transformers==2.7.0
transformers==4.45.1
accelerate==0.27.2

# ---------------- FAISS ANN Search ----------------
faiss-cpu==1.7.4

# ---------------- PDF Utils (optional) ----------------
pypdf==4.1.0
